PUT and DELETE Request Mapping

+++ src/main/java/com/vinodh/dao/UserRegistrationDAO.java
@@ -17,7 +17,7 @@ public interface UserRegistrationDAO {
 
 	List<ApplicationUser> loadAllUserDetails();
 
-	int deleteUser(int userId);
+	int deleteUser(Long userId);
 
 	void updateUser(ApplicationUser user);
